third_party/blink/web_tests/wpt_internal/bluetooth/requestDevice/name-empty-filter.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(t => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
      .then(
          () => promise_rejects_dom(
              t, 'NotFoundError',
              requestDeviceWithTrustedClick(
                  {filters: [{name: '', services: ['heart_rate']}]})));
}, 'A named device is not matched by a filter with an empty name.');
</script>
